Update chainlink pallet unit tests
