+++ pallet-chainlink/src/lib.rs
@@ -23,6 +23,9 @@ use frame_support::traits::{Get, ReservableCurrency, Currency, BalanceStatus, Un
 use sp_std::prelude::*;
 use frame_system::ensure_signed;
 
+mod tests;
+
+
 // A trait allowing to inject Operator results back into the specified Call
 pub trait CallbackWithParameter {
 	fn with_result(&self, result: Vec<u8>) -> Option<Self> where Self: core::marker::Sized;
@@ -222,224 +225,4 @@ decl_module! {
 		}
 
 	}
-}
-
-#[cfg(test)]
-mod tests {
-	use super::*;
-
-	use codec::{Decode, Encode};
-	use frame_support::{impl_outer_event, impl_outer_origin, parameter_types, weights::Weight};
-	use sp_core::H256;
-	// The testing primitives are very useful for avoiding having to work with signatures
-	// or public keys. `u64` is used as the `AccountId` and no `Signature`s are required.
-	use sp_runtime::{
-		Perbill,
-		testing::Header,
-		traits::{BlakeTwo256, OnFinalize, IdentityLookup},
-	};
-	use frame_system::{EventRecord, Phase};
-
-	impl_outer_origin! {
-		pub enum Origin for Runtime {}
-	}
-
-	#[derive(Clone, Eq, PartialEq, Debug)]
-	pub struct Runtime;
-	parameter_types! {
-		pub const BlockHashCount: u64 = 250;
-		pub const MaximumBlockWeight: Weight = 1024;
-		pub const MaximumBlockLength: u32 = 2 * 1024;
-		pub const AvailableBlockRatio: Perbill = Perbill::one();
-	}
-	impl frame_system::Trait for Runtime {
-		type Origin = Origin;
-		type Index = u64;
-		type BlockNumber = u64;
-		type Hash = H256;
-		type Call = ();
-		type Hashing = BlakeTwo256;
-		type AccountId = u64;
-		type Lookup = IdentityLookup<Self::AccountId>;
-		type Header = Header;
-		type Event = TestEvent;
-		type BlockHashCount = BlockHashCount;
-		type MaximumBlockWeight = MaximumBlockWeight;
-		type MaximumBlockLength = MaximumBlockLength;
-		type AvailableBlockRatio = AvailableBlockRatio;
-		type Version = ();
-		type ModuleToIndex = ();
-	}
-	parameter_types! {
-		pub const ExistentialDeposit: u64 = 1;
-		pub const TransferFee: u64 = 0;
-		pub const CreationFee: u64 = 0;
-	}
-	impl balances::Trait for Runtime {
-		type Balance = u64;
-		type OnFreeBalanceZero = ();
-		type OnNewAccount = ();
-		type DustRemoval = ();
-		type TransferPayment = ();
-		type Event = TestEvent;
-		type ExistentialDeposit = ExistentialDeposit;
-		type TransferFee = TransferFee;
-		type CreationFee = CreationFee;
-	}
-	impl Trait for Runtime {
-		type Event = TestEvent;
-		type Currency = balances::Module<Runtime>;
-		type Callback = module2::Call<Runtime>;
-		type ValidityPeriod = ValidityPeriod;
-	}
-	impl module2::Trait for Runtime {
-	}
-	parameter_types! {
-		pub const ValidityPeriod: u64 = 10;
-	}
-
-	mod chainlink {
-		pub use crate::Event;
-	}
-
-	impl_outer_event! {
-		pub enum TestEvent for Runtime {
-			balances<T>,
-			chainlink<T>,
-		}
-	}
-
-	type System = frame_system::Module<Runtime>;
-
-	fn new_test_ext() -> sp_io::TestExternalities {
-		let mut t = frame_system::GenesisConfig::default().build_storage::<Runtime>().unwrap();
-		balances::GenesisConfig::<Runtime>{
-			balances: vec![(1, 10), (2, 20)],
-			vesting: vec![],
-		}.assimilate_storage(&mut t).unwrap();
-		balances::GenesisConfig::<Runtime>::default().assimilate_storage(&mut t).unwrap();
-		t.into()
-	}
-
-	mod module2 {
-		use super::*;
-
-		pub trait Trait: frame_system::Trait {}
-
-		frame_support::decl_module! {
-			pub struct Module<T: Trait> for enum Call
-				where origin: <T as frame_system::Trait>::Origin
-			{
-				pub fn callback(_origin, result: Vec<u8>) -> frame_support::dispatch::DispatchResult {
-					let r : u128 = u128::decode(&mut &result[..]).map_err(|err| err.what())?;
-					<Result>::put(r);
-					Ok(())
-				}
-			}
-		}
-
-		frame_support::decl_storage! {
-			trait Store for Module<T: Trait> as TestStorage {
-				pub Result: u128;
-			}
-		}
-
-		impl <T: Trait> CallbackWithParameter for Call<T> {
-			fn with_result(&self, result: Vec<u8>) -> Option<Self> {
-				match *self {
-					Call::callback(_) => Some(Call::callback(result)),
-					_ => None
-				}
-			}
-		}
-
-	}
-
-	#[test]
-	fn operators_can_be_registered() {
-		new_test_ext().execute_with(|| {
-			assert!(!<Operators<Runtime>>::get(1));
-			assert!(<Module<Runtime>>::register_operator(Origin::signed(1)).is_ok());
-			assert!(<Operators<Runtime>>::get(1));
-			assert!(<Module<Runtime>>::unregister_operator(Origin::signed(1)).is_ok());
-			assert!(!<Operators<Runtime>>::get(1));
-		});
-
-		new_test_ext().execute_with(|| {
-			assert!(<Module<Runtime>>::unregister_operator(Origin::signed(1)).is_err());
-			assert!(!<Operators<Runtime>>::get(1));
-		});
-
-	}
-
-	#[test]
-	fn initiate_requests() {
-
-		new_test_ext().execute_with(|| {
-			assert!(<Module<Runtime>>::register_operator(Origin::signed(1)).is_ok());
-			assert!(<Module<Runtime>>::initiate_request(Origin::signed(2), 1, vec![], 1, vec![], 0, module2::Call::<Runtime>::callback(vec![]).into()).is_err());
-		});
-
-		new_test_ext().execute_with(|| {
-			assert!(<Module<Runtime>>::initiate_request(Origin::signed(2), 1, vec![], 1, vec![], 1, module2::Call::<Runtime>::callback(vec![]).into()).is_err());
-		});
-
-		new_test_ext().execute_with(|| {
-			assert!(<Module<Runtime>>::register_operator(Origin::signed(1)).is_ok());
-			assert!(<Module<Runtime>>::initiate_request(Origin::signed(2), 1, vec![], 1, vec![], 2, module2::Call::<Runtime>::callback(vec![]).into()).is_ok());
-			assert!(<Module<Runtime>>::callback(Origin::signed(3), 0, 10.encode()).is_err());
-		});
-
-		new_test_ext().execute_with(|| {
-			assert!(<Module<Runtime>>::callback(Origin::signed(1), 0, 10.encode()).is_err());
-		});
-
-		new_test_ext().execute_with(|| {
-			assert!(<Module<Runtime>>::register_operator(Origin::signed(1)).is_ok());
-
-			assert_eq!(
-				*System::events().last().unwrap(),
-				EventRecord {
-					phase: Phase::ApplyExtrinsic(0),
-					event: TestEvent::chainlink(RawEvent::OperatorRegistered(1)),
-					topics: vec![],
-				}
-			);
-
-			let parameters = ("a", "b");
-			let data = parameters.encode();
-			assert!(<Module<Runtime>>::initiate_request(Origin::signed(2), 1, vec![], 1, data.clone(), 2, module2::Call::<Runtime>::callback(vec![]).into()).is_ok());
-
-			assert_eq!(
-				*System::events().last().unwrap(),
-				EventRecord {
-					phase: Phase::ApplyExtrinsic(0),
-					event: TestEvent::chainlink(RawEvent::OracleRequest(1, vec![], 0, 2, 1, data.clone(), "Chainlink.callback".into(), 2)),
-					topics: vec![],
-				}
-			);
-
-			let r = <(Vec<u8>, Vec<u8>)>::decode(&mut &data[..]).unwrap().0;
-			assert_eq!("a", std::str::from_utf8(&r).unwrap());
-
-			let result = 10;
-			assert!(<Module<Runtime>>::callback(Origin::signed(1), 0, result.encode()).is_ok());
-			assert_eq!(module2::Result::get(), result);
-		});
-
-	}
-
-	#[test]
-	pub fn on_finalize() {
-
-		new_test_ext().execute_with(|| {
-			assert!(<Module<Runtime>>::register_operator(Origin::signed(1)).is_ok());
-			assert!(<Module<Runtime>>::initiate_request(Origin::signed(2), 1, vec![], 1, vec![], 2, module2::Call::<Runtime>::callback(vec![]).into()).is_ok());
-			<Module<Runtime> as OnFinalize<u64>>::on_finalize(20);
-			// Request has been killed, too old
-			assert!(<Module<Runtime>>::callback(Origin::signed(1), 0, 10.encode()).is_err());
-		});
-
-	}
-
 }
